+++ CHANGELOG.md
@@ -10,7 +10,7 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 ### Adds
 
-- Adds the `Filtering` page to the `Style Guide`.
+- Adds the `Filtering` subsection to the `Style Guide`.
 
 ## 3.5.1 (December 19, 2024)